[REF] hr_work_entry_holidays,hr_holidays_contract: create bridge module between hr_holidays and hr_contract
Before this commit, there is no module between hr_holidays and hr_contracts if
both module are installed so time off duration computation is not correct if we
dont use work entry.
In this commit, moved leave and contract related code from
hr_work_entry_holidays to a new module hr_holidays_contract which work as bridge
module between hr_holiday and hr_contracts.

task-3462279
