+++ addons/hr_work_entry_holidays/__manifest__.py
@@ -14,7 +14,7 @@ Manage Time Off in Payslips
 
 This application allows you to integrate time off in payslips.
     """,
-    'depends': ['hr_holidays', 'hr_work_entry_contract'],
+    'depends': ['hr_holidays', 'hr_holidays_contract', 'hr_work_entry_contract'],
     'data': [
         'data/hr_payroll_holidays_data.xml',
         'views/hr_leave_views.xml',
