+++ addons/hr_work_entry_holidays/models/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from . import hr_contract
-from . import hr_employee_base
 from . import hr_leave
 from . import hr_work_entry
