+++ addons/hr_work_entry_holidays/views/hr_leave_views.xml
@@ -16,16 +16,4 @@
         </field>
     </record>
 
-    <record id="hr_leave_view_tree_inherit_work_entry" model="ir.ui.view">
-        <field name="name">hr.holidays.view.tree.inherit.work.entry</field>
-        <field name="model">hr.leave</field>
-        <field name="inherit_id" ref="hr_holidays.hr_leave_view_tree"/>
-        <field name="mode">primary</field>
-        <field name="arch" type="xml">
-            <xpath expr="//tree" position="attributes">
-                <attribute name="default_order">date_from</attribute>
-            </xpath>
-        </field>
-    </record>
-
 </odoo>
